Some progress on OpenGL. Milestones 1 and 2 complete.
I've given up on the picture-in-picture preview, because GLKViews are not nearly as easy to use as the documentation would have you believe, and I'm tired. I do have a ViewController which can render to a GL context working, though, so it should be easy enough to add the preview stuff there. This commit is my submission for the first milestone, with one caveat: as I can't test on a multitouch device, I have no clue whether or not the multitouch gestures work, so pinch-to-zoom-and-rotate is a tentative inclusion.