A shameful amount of changes.
Lots of progress since the last commit. Spine manipulation on cylinders. Radius manipulation on cylinders. Improvements on end cap generation for both shapes. Note that this commit takes out a lot of the view manipulation, because I'm going to switch from using UIGestureRecognizers to rolling my own -- this is required for doing TRS all at once for shapes and TS for the view. Next step is this.