Same-* annotations, out-of-plane tilt.
Lots of cool stuff in this commit. New handles for manipulating tilt at any point along the spine of a cylinderoid, with linear interpolation along points that don't have tilts defined. Same-scale annotation, which forces two radii to be the same. Same-length annotation, which forces two cylinderoids to be the same length.