Lots of changes, all for the connect annotation.
It has a terrible, awful, very-bad user interface right now, but I'm going to overhaul the annotation interface in the next milestone (?) anyway, so so be it. Select Connect tool, click once to select the first shape (nothing happens), click a second time to select the second shape (it turns orange), click a third time to pick the connection location (it stops being orange). Connection points are drawn as green handles when a shape is selected.